Moves OldDoFn to runners-core
