+++ runners/google-cloud-dataflow-java/src/main/java/org/apache/beam/runners/dataflow/util/DoFnInfo.java
@@ -21,7 +21,6 @@ import java.io.Serializable;
 import java.util.Map;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.transforms.DoFn;
-import org.apache.beam.sdk.transforms.OldDoFn;
 import org.apache.beam.sdk.util.WindowingStrategy;
 import org.apache.beam.sdk.values.PCollectionView;
 import org.apache.beam.sdk.values.TupleTag;
@@ -41,8 +40,7 @@ public class DoFnInfo<InputT, OutputT> implements Serializable {
   private final Map<Long, TupleTag<?>> outputMap;
 
   /**
-   * Creates a {@link DoFnInfo} for the given {@link Serializable} object, which is expected to be a
-   * {@link DoFn} or {@link OldDoFn} or other context-appropriate UDF blob.
+   * Creates a {@link DoFnInfo} for the given {@link DoFn}.
    */
   public static <InputT, OutputT> DoFnInfo<InputT, OutputT> forFn(
       DoFn<InputT, OutputT> doFn,
